rewrite nmt without attention.
